Add quirks controls, expose more c stubs
The quirks are process-scope (globals in the .so) with all the issues
that comes with.  Significant effort was spent trying to avoid this but
the Tcl core's objtype handling makes this extremely difficult (no
interp context for the update_string_rep callbacks) and the result was
still broken and much uglier.